requirements.txt
fastapi[standard]
PyJWT
uvicorn
python-dotenv
motor
dnspython